requirements.txt
argon2-cffi>=21.3.0
Flask>=2.2.5
itsdangerous>=2.0.0
Jinja2>=3.0.0
markdown
markdown-extensions
markdown-it-py>=3.0.0
MarkupSafe>=2.0.0
mdit-py-plugins>=0.4.0
pymdown-extensions
python-frontmatter>=1.0.0
python-markdown-math
PyYAML>=6.0.0
rfeed
Werkzeug==2.2.2
